chore(autoware_adapi_specs): rename ad_api to adapi
