+++ common/autoware_adapi_specs/CHANGELOG.rst
