+++ common/autoware_adapi_specs/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.14)
-project(autoware_ad_api_specs)
+project(autoware_adapi_specs)
 
 find_package(autoware_cmake REQUIRED)
 autoware_package()
