+++ common/autoware_adapi_specs/include/autoware/adapi_specs/localization.hpp
@@ -12,25 +12,32 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef AUTOWARE__AD_API_SPECS__FAIL_SAFE_HPP_
-#define AUTOWARE__AD_API_SPECS__FAIL_SAFE_HPP_
+#ifndef AUTOWARE__AD_API_SPECS__LOCALIZATION_HPP_
+#define AUTOWARE__AD_API_SPECS__LOCALIZATION_HPP_
 
 #include <rclcpp/qos.hpp>
 
-#include <autoware_adapi_v1_msgs/msg/mrm_state.hpp>
+#include <autoware_adapi_v1_msgs/msg/localization_initialization_state.hpp>
+#include <autoware_adapi_v1_msgs/srv/initialize_localization.hpp>
 
-namespace autoware::ad_api_specs::fail_safe
+namespace autoware::adapi_specs::localization
 {
 
-struct MrmState
+struct Initialize
 {
-  using Message = autoware_adapi_v1_msgs::msg::MrmState;
-  static constexpr char name[] = "/api/fail_safe/mrm_state";
+  using Service = autoware_adapi_v1_msgs::srv::InitializeLocalization;
+  static constexpr char name[] = "/api/localization/initialize";
+};
+
+struct InitializationState
+{
+  using Message = autoware_adapi_v1_msgs::msg::LocalizationInitializationState;
+  static constexpr char name[] = "/api/localization/initialization_state";
   static constexpr size_t depth = 1;
   static constexpr auto reliability = RMW_QOS_POLICY_RELIABILITY_RELIABLE;
   static constexpr auto durability = RMW_QOS_POLICY_DURABILITY_TRANSIENT_LOCAL;
 };
 
-}  // namespace autoware::ad_api_specs::fail_safe
+}  // namespace autoware::adapi_specs::localization
 
-#endif  // AUTOWARE__AD_API_SPECS__FAIL_SAFE_HPP_
+#endif  // AUTOWARE__AD_API_SPECS__LOCALIZATION_HPP_
