+++ common/autoware_adapi_specs/include/autoware/adapi_specs/motion.hpp
@@ -12,32 +12,32 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef AUTOWARE__AD_API_SPECS__LOCALIZATION_HPP_
-#define AUTOWARE__AD_API_SPECS__LOCALIZATION_HPP_
+#ifndef AUTOWARE__AD_API_SPECS__MOTION_HPP_
+#define AUTOWARE__AD_API_SPECS__MOTION_HPP_
 
 #include <rclcpp/qos.hpp>
 
-#include <autoware_adapi_v1_msgs/msg/localization_initialization_state.hpp>
-#include <autoware_adapi_v1_msgs/srv/initialize_localization.hpp>
+#include <autoware_adapi_v1_msgs/msg/motion_state.hpp>
+#include <autoware_adapi_v1_msgs/srv/accept_start.hpp>
 
-namespace autoware::ad_api_specs::localization
+namespace autoware::adapi_specs::motion
 {
 
-struct Initialize
+struct AcceptStart
 {
-  using Service = autoware_adapi_v1_msgs::srv::InitializeLocalization;
-  static constexpr char name[] = "/api/localization/initialize";
+  using Service = autoware_adapi_v1_msgs::srv::AcceptStart;
+  static constexpr char name[] = "/api/motion/accept_start";
 };
 
-struct InitializationState
+struct State
 {
-  using Message = autoware_adapi_v1_msgs::msg::LocalizationInitializationState;
-  static constexpr char name[] = "/api/localization/initialization_state";
+  using Message = autoware_adapi_v1_msgs::msg::MotionState;
+  static constexpr char name[] = "/api/motion/state";
   static constexpr size_t depth = 1;
   static constexpr auto reliability = RMW_QOS_POLICY_RELIABILITY_RELIABLE;
   static constexpr auto durability = RMW_QOS_POLICY_DURABILITY_TRANSIENT_LOCAL;
 };
 
-}  // namespace autoware::ad_api_specs::localization
+}  // namespace autoware::adapi_specs::motion
 
-#endif  // AUTOWARE__AD_API_SPECS__LOCALIZATION_HPP_
+#endif  // AUTOWARE__AD_API_SPECS__MOTION_HPP_
