+++ common/autoware_adapi_specs/include/autoware/adapi_specs/routing.hpp
@@ -23,7 +23,7 @@
 #include <autoware_adapi_v1_msgs/srv/set_route.hpp>
 #include <autoware_adapi_v1_msgs/srv/set_route_points.hpp>
 
-namespace autoware::ad_api_specs::routing
+namespace autoware::adapi_specs::routing
 {
 
 struct SetRoutePoints
@@ -74,6 +74,6 @@ struct Route
   static constexpr auto durability = RMW_QOS_POLICY_DURABILITY_TRANSIENT_LOCAL;
 };
 
-}  // namespace autoware::ad_api_specs::routing
+}  // namespace autoware::adapi_specs::routing
 
 #endif  // AUTOWARE__AD_API_SPECS__ROUTING_HPP_
